# Cache transformations within a phase