--- a/internal/corazawaf/transaction.go
+++ b/internal/corazawaf/transaction.go
@@ -104,6 +104,8 @@ type Transaction struct {
 	audit bool
 
 	variables TransactionVariables
+
+	transformationCache map[transformationKey]transformationValue
 }
 
 func (tx *Transaction) ID() string {