--- a/internal/corazawaf/waf.go
+++ b/internal/corazawaf/waf.go
@@ -190,6 +190,7 @@ func (w *WAF) newTransactionWithID(id string) *Transaction {
 			MemoryLimit: w.RequestBodyInMemoryLimit,
 		})
 		tx.variables = *NewTransactionVariables()
+		tx.transformationCache = map[transformationKey]transformationValue{}
 	}
 
 	// set capture variables