We currently compute the same transformation per variable many many times per phase. This adds a cache to reduce this repetitive computation. It notably reduces allocations by half in the CRS benchmarks, which has some performance impact here and will have bigger performance impact where GC is slower like wasm.

After
```
BenchmarkCRSLargePOST
BenchmarkCRSLargePOST-10    	      12	  95202174 ns/op	 3631520 B/op	   45437 allocs/op
BenchmarkCRSSimplePOST
BenchmarkCRSSimplePOST-10    	     631	   1887620 ns/op	 1294996 B/op	   45180 allocs/op
```

Before
```
BenchmarkCRSLargePOST
BenchmarkCRSLargePOST-10    	      12	  97733875 ns/op	 7036534 B/op	   46887 allocs/op
BenchmarkCRSSimplePOST
BenchmarkCRSSimplePOST-10    	     631	   1883263 ns/op	 1302137 B/op
```